pyproject.toml
[project]
name = "idp-service"
version = "0.1.0"
description = "Enterprise Identity Provider Microservice"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyjwt[crypto]>=2.8.0",
    "passlib[argon2]>=1.7.4",
    "pyotp>=2.9.0",
    "cryptography>=41.0.0",
    "alembic>=1.12.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "python-json-logger>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pytest-cov>=4.1.0",
    "fakeredis>=2.0.0"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
